add confirmation dialog to the deactivate workspace endpoint

+++ src/components/containers/workspace/WorkspaceDetails.jsx
@@ -3,38 +3,46 @@ import {
   deleteWorkspace,
   updateWorkspace,
 } from '@/app/_actions/config-actions'
+import PromptModal from '@/components/base/Prompt'
+import Prompt from '@/components/base/Prompt'
 import { Button } from '@/components/ui/Button'
 import { Input } from '@/components/ui/InputField'
 import useWorkspaces from '@/hooks/useWorkspace'
 import { SETUP_QUERY_KEY, WORKSPACES_QUERY_KEY } from '@/lib/constants'
 import { cn, notify } from '@/lib/utils'
-import { Switch } from '@nextui-org/react'
+import { Switch, useDisclosure } from '@nextui-org/react'
 import { useQueryClient } from '@tanstack/react-query'
 import { useRouter } from 'next/navigation'
 import React, { useEffect, useState } from 'react'
 
-const INITIAL_WORKSPACE = {
-  workspace: '',
-  description: '',
-}
-
 function WorkspaceDetails({ WorkSpaceID }) {
+  const { back } = useRouter()
+  const queryClient = useQueryClient()
+  const [loading, setLoading] = useState(false)
+  const { isOpen, onOpen, onClose } = useDisclosure()
+  const [deleteError, setDeleteError] = useState({ status: false, message: '' })
+  const [deleteLoading, setDeleteLoading] = useState(false)
+  const [deleteWorkspaceName, setDeleteWorkspaceName] = useState('')
   const { workspaces, allWorkspaces } = useWorkspaces()
   const workspace = allWorkspaces.find(
     (workspace) => workspace.ID === WorkSpaceID,
   )
 
-  const { back } = useRouter()
-  const queryClient = useQueryClient()
-  const [loading, setLoading] = useState(false)
-  const [deleteLoading, setDeleteLoading] = useState(false) // DEACTIVATION STATE
   const [isSandbox, setIsSandbox] = useState(
     workspace?.workspace.toLowerCase() == 'sandbox',
   )
 
-  // TO EDIT THE WORKSPACE FIELDS
-  const [newWorkspace, setNewWorkspace] = useState(INITIAL_WORKSPACE)
+  const [newWorkspace, setNewWorkspace] = useState({
+    workspace: workspace?.workspace,
+    description: workspace?.description,
+  })
+
   const [isVisible, setIsVisible] = useState(workspace?.isVisible)
+
+  const noChangesToSave =
+    newWorkspace.workspace == workspace?.workspace &&
+    newWorkspace.description == workspace?.description
+
   function editWorkspaceField(fields) {
     setNewWorkspace((prev) => {
       return { ...prev, ...fields }
@@ -59,8 +67,9 @@ function WorkspaceDetails({ WorkSpaceID }) {
 
     if (response.success) {
       queryClient.invalidateQueries({ queryKey: [SETUP_QUERY_KEY] })
+      queryClient.invalidateQueries({ queryKey: [WORKSPACES_QUERY_KEY] })
       setLoading(false)
-      notify('success', 'Workspaces created successfully!')
+      notify('success', 'Changes Saved!')
       return
     }
 
@@ -70,6 +79,16 @@ function WorkspaceDetails({ WorkSpaceID }) {
 
   async function handleDeleteWorkspace() {
     setDeleteLoading(true)
+
+    if (deleteWorkspaceName !== workspace?.workspace) {
+      setDeleteLoading(false)
+      setDeleteError({
+        status: true,
+        message: 'Type the workspace name to confirm delete',
+      })
+
+      return
+    }
 
     const response = await deleteWorkspace(WorkSpaceID)
 
@@ -118,24 +137,19 @@ function WorkspaceDetails({ WorkSpaceID }) {
     }
   }, [])
 
+  // CLEAR ERROR STATE
+  useEffect(() => {
+    setDeleteError({ status: false, message: '' })
+
+    return () => {
+      setDeleteError({ status: false, message: '' })
+    }
+  }, [deleteWorkspaceName])
+
   return (
     <>
       <div className="flex w-full flex-1 flex-col gap-4">
         <div className="flex flex-col gap-6 ">
-          {/* <div className="flex w-full items-center gap-x-8">
-            <img
-              alt=""
-              src="https://images.unsplash.com/photo-1499951360447-b19be8fe80f5?q=80&w=1770&auto=format&fit=crop&ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8fA%3D%3D"
-              className="h-24 w-24 flex-none rounded-lg bg-gray-800 object-cover"
-            />
-            <div>
-              <Button type="button">Change Workspace Image</Button>
-              <p className="mt-2 text-xs leading-5 text-gray-400">
-                JPG, GIF or PNG. 2MB max.
-              </p>
-            </div>
-          </div> */}
-
           <form
             onSubmit={handleUpdateWorkspace}
             role={'edit-workspace-details'}
@@ -162,9 +176,9 @@ function WorkspaceDetails({ WorkSpaceID }) {
             {!isSandbox && (
               <Button
                 type="submit"
-                isDisabled={loading || isSandbox}
+                isDisabled={loading || isSandbox || noChangesToSave}
                 isLoading={loading}
-                loadingText={'Saving... '}
+                loadingText={'Saving...'}
               >
                 Save Changes
               </Button>
@@ -193,25 +207,56 @@ function WorkspaceDetails({ WorkSpaceID }) {
             Deactivate Workspace
           </h2>
           <p className="text-sm leading-6 text-gray-400">
-            If you wish to deactivate or delete your workspace, you can do so
-            here. This action can only be reversed by contacting our support
-            team. Please note that all information related to this workspace
-            will be retained for audit purposes in accordance with regulatory
-            requirements.
+            If you wish to deactivate or delete your workspace. This action can
+            only be reversed by contacting our support team. Please note that
+            all information related to this workspace will be retained for audit
+            purposes in accordance with regulatory requirements.
           </p>
         </div>
 
         <Button
           color="danger"
           type="submit"
-          isDisabled={deleteLoading}
-          isLoading={deleteLoading}
-          onClick={handleDeleteWorkspace}
+          onClick={onOpen}
           className="self-end rounded-md px-3 py-2 text-sm font-semibold text-white shadow-sm hover:bg-red-400"
         >
           Yes, deactivate my workspace
         </Button>
       </div>
+
+      {/* MODALS */}
+      {isOpen && (
+        <PromptModal
+          isOpen={isOpen}
+          onOpen={onOpen}
+          onClose={onClose}
+          title="Delete/Deactivate Workspace"
+          onConfirm={handleDeleteWorkspace}
+          confirmText="Deactivate"
+          isDisabled={deleteLoading}
+          isLoading={deleteLoading}
+          isDismissable={false}
+        >
+          <p className="leading-2 m-0">
+            <strong>Are you sure you want to perform this action?</strong>
+          </p>
+          <p className="text-sm text-slate-700">
+            This action cannot be undone. Please type{' '}
+            <code className="rounded-md bg-primary/10 p-1 px-2 font-medium text-primary-700">
+              {workspace?.workspace}
+            </code>{' '}
+            to confirm your choice to proceed.
+          </p>
+
+          <Input
+            label="Confirm Delete"
+            isDisabled={deleteLoading}
+            onError={deleteError.status}
+            errorText={deleteError.message}
+            onChange={(e) => setDeleteWorkspaceName(e.target.value)}
+          />
+        </PromptModal>
+      )}
     </>
   )
 }
